lunch-api/requirements.txt
django==3.0.6
djangorestframework==3.11.0
django-cors-headers

# dev packages
factory-boy